feat(playground): harden GenUI catalog normalization and conversation lifecycle
Add catalog data normalizer to coerce AI payload types, strip injected
component fields from schemas, and make conversation startup async with
request queuing and session epoch guards. Update deck tools, prompts,
and tests with shared fake agent client helper.

+++ packages/playground/lib/features/ai/core/ai/catalog/remix_component_renderer.dart
@@ -181,7 +181,7 @@ class ComponentTreeRenderer extends StatelessWidget {
   }
 
   Widget _buildProgress(UiNodeType node) {
-    final value = (node.value ?? 0.5).clamp(0.0, 1.0);
+    final value = (node.value ?? 0.5).clamp(0.0, 1.0).toDouble();
     final label = node.label;
 
     final row = FlexBoxStyler()
@@ -281,7 +281,7 @@ class ComponentTreeRenderer extends StatelessWidget {
   }
 
   Widget _buildSlider(UiNodeType node) {
-    final value = (node.value ?? 0.5).clamp(0.0, 1.0);
+    final value = (node.value ?? 0.5).clamp(0.0, 1.0).toDouble();
 
     final col = FlexBoxStyler()
         .column()
